refactor(ai): re-express the chat-op grammar on the ops layer
Issue #1764, step 2: the chat grammar — arg models, op registry,
parse_tool_call — moves verbatim from src/ai/chat_ops.py to
src/ops/grammar.py, beside the operation registry that maps each chat op
onto its canonical executor. src/ai/chat_ops.py becomes a re-export shim,
so every importer (src.ai.chat, the op-contract tests, the web contract's
validation surface) sees byte-identical names, schemas, and errors.

The chat system prompt now takes its device line from the shared teaching
module (src.ops.teaching.dimensions_phrase) — same rendered text, one
source for both surfaces once the MCP side switches in step 3.

Part of #1849.
